+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 * Cleaner state saving code and compress with stream instead of packets
 * Queue saving memory pages for less write() calls
 * Queue saving compressed memory pages to be saved in temporary memory segment
+* Change ImGui font to Roboto Medium
 
 ### Fixed